Nifi-1540 - Updated name, displayname, comments on provisioning for IT, moved tests to stream package, based on comments
